requirements.txt
# Core RAG dependencies
anthropic>=0.34.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0

# Document processing
markdown>=3.6
python-frontmatter>=1.1.0

# API & Web interface
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# CLI interface
rich>=13.7.0
typer>=0.12.0

# Configuration & utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
